+++ util.c
@@ -42,7 +42,7 @@ char *d_dup_pathname_str(const char *s) {
 }
 
 char *dup_str(const char *s, const char *e) {
-  int l = e ? e - s : strlen(s);
+  size_t l = e ? (size_t) (e - s) : strlen(s);
   char *ss = (char *)MALLOC(l + 1);
   memcpy(ss, s, l);
   ss[l] = 0;
